Add JWT authentication and authorization to CourseApi

+++ course-service/learning/src/main/java/com/online/api/CourseApi.java
@@ -2,10 +2,12 @@ package com.online.api;
 
 import com.online.controllers.CourseRepository;
 import com.online.model.Course;
+import com.online.service.JwtParser;
 
 import jakarta.ejb.EJB;
 import jakarta.ejb.Stateless;
 import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.CookieParam;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
@@ -15,7 +17,10 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+
 import java.util.List;
+
+import org.jose4j.jwt.JwtClaims;
 
 @Path("/course")
 @Consumes(MediaType.APPLICATION_JSON)
@@ -25,6 +30,34 @@ public class CourseApi {
 
   @EJB
   CourseRepository repository;
+
+  JwtParser jwtParser = new JwtParser();
+
+  @GET
+  @Path("/test")
+  public Response test(@CookieParam("jwt") String jwt) {
+    try {
+      System.out.println(jwt);
+
+      if (jwt == null) {
+        return Response.status(Response.Status.UNAUTHORIZED).build();
+      }
+
+      JwtClaims claims = jwtParser.parseClaims(jwt);
+      if (claims == null) {
+        System.out.println("Claims are null");
+        return Response.status(Response.Status.UNAUTHORIZED).build();
+      }
+
+      System.out.println("ID:    " + claims.getClaimValue("id"));
+      System.out.println("ROLE:  " + claims.getClaimValue("role"));
+
+      return Response.ok("Test").build();
+    } catch (Exception e) {
+      e.printStackTrace();
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+  }
 
   @POST
   @Path("/create")
@@ -58,7 +91,21 @@ public class CourseApi {
 
   @PUT
   @Path("/update/{id}")
-  public Response updateCourse(@PathParam("id") long id, Course course) {
+  public Response updateCourse(@CookieParam("jwt") String jwt, @PathParam("id") long id, Course course) {
+    if (jwt == null) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    JwtClaims claims = jwtParser.parseClaims(jwt);
+    if (claims == null) {
+      System.out.println("Claims are null");
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    if (!claims.getClaimValue("role").equals("instructor".toLowerCase())) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
     String response = repository.updateCourse(id, course);
     if (response == null) {
       return Response.serverError().build();
@@ -68,7 +115,21 @@ public class CourseApi {
 
   @PUT
   @Path("/accept/{id}")
-  public Response acceptCourseById(@PathParam("id") long id) {
+  public Response acceptCourseById(@CookieParam("jwt") String jwt, @PathParam("id") long id) {
+    if (jwt == null) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    JwtClaims claims = jwtParser.parseClaims(jwt);
+    if (claims == null) {
+      System.out.println("Claims are null");
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    if (!claims.getClaimValue("role").equals("instructor".toLowerCase())) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
     String response = repository.acceptCourse(id);
     if (response == null) {
       return Response.serverError().build();
@@ -78,7 +139,22 @@ public class CourseApi {
 
   @PUT
   @Path("/reject/{id}")
-  public Response rejectCourseById(@PathParam("id") long id) {
+  public Response rejectCourseById(@CookieParam("jwt") String jwt, @PathParam("id") long id) {
+
+    if (jwt == null) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    JwtClaims claims = jwtParser.parseClaims(jwt);
+    if (claims == null) {
+      System.out.println("Claims are null");
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    if (!claims.getClaimValue("role").equals("instructor".toLowerCase())) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
     String response = repository.rejectCourse(id);
     if (response == null) {
       return Response.serverError().build();
@@ -88,7 +164,21 @@ public class CourseApi {
 
   @DELETE
   @Path("/remove/{id}")
-  public Response removeCourseById(@PathParam("id") long id) {
+  public Response removeCourseById(@CookieParam("jwt") String jwt, @PathParam("id") long id) {
+    if (jwt == null) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    JwtClaims claims = jwtParser.parseClaims(jwt);
+    if (claims == null) {
+      System.out.println("Claims are null");
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
+    if (!claims.getClaimValue("role").equals("instructor".toLowerCase())) {
+      return Response.status(Response.Status.UNAUTHORIZED).build();
+    }
+
     String response = repository.removeCourse(id);
     if (response == null) {
       return Response.serverError().build();
@@ -105,7 +195,5 @@ public class CourseApi {
     }
     return Response.ok(result).build();
   }
-
-
 
 }
